+++ HISTORY.rst
@@ -6,3 +6,17 @@ History
 ------------------
 
 * First working release.*
+
+1.148 (11-07-2019)
+
+------------------
+
+* Fixed some setup and requirement related Docs. *
+
+
+1.247 (14-07-2019)
+
+* Added Brightness/visiblity info for webcam.
+  You can see the visiblity condition ranging
+  from very poor to excellent live in webcam 
+  frame displayed. *